Moved existing tests to new tests folder
