+++ src/game_systems/Cargo.toml
@@ -18,5 +18,14 @@ temper-commands = { workspace = true }
 temper-config = { workspace = true }
 temper-scheduler = { workspace = true }
 
+[dev-dependencies]
+bevy_math = { workspace = true }
+temper-components = { workspace = true }
+temper-core = { workspace = true }
+temper-entities = { workspace = true }
+temper-macros = { workspace = true }
+temper-messages = { workspace = true }
+temper-state = { workspace = true }
+
 [lints]
 workspace = true
