+++ src/game_systems/src/physics/src/gravity.rs
@@ -51,188 +51,3 @@ pub fn handle(mut entities: EntityQuery, state: Res<GlobalStateResource>) {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use bevy_ecs::prelude::*;
-    use bevy_math::DVec3;
-    use bevy_math::Vec3A;
-    use temper_components::player::grounded::OnGround;
-    use temper_components::player::velocity::Velocity;
-    use temper_entities::markers::HasGravity;
-    use temper_macros::block;
-    use temper_state::create_test_state;
-
-    /// Creates a chunk with water blocks at the specified positions
-    /// This helper function is used to set up test scenarios where entities are in water
-    fn create_chunk_with_water(state: &GlobalStateResource, chunk_pos: ChunkPos) {
-        // Load or generate the chunk
-        let mut chunk = state
-            .0
-            .world
-            .get_or_generate_mut(chunk_pos, Dimension::Overworld)
-            .expect("Failed to load or generate chunk");
-
-        chunk.fill(block!("water", { level: 0 }));
-    }
-
-    #[test]
-    fn test_gravity_application() {
-        let mut world = World::new();
-        let (state, _temp_dir) = create_test_state();
-        world.insert_resource(state);
-
-        let entity = world
-            .spawn((
-                Velocity { vec: Vec3A::ZERO },
-                OnGround(false),
-                Position {
-                    coords: DVec3::new(0.0, 100.0, 0.0),
-                },
-                HasGravity,
-            ))
-            .id();
-
-        let mut schedule = Schedule::default();
-        schedule.add_systems(handle);
-
-        // Run the gravity system
-        schedule.run(&mut world);
-
-        let vel = world.get::<Velocity>(entity).unwrap();
-        assert!(
-            vel.vec.y < 0.0,
-            "Velocity Y should be negative after gravity application"
-        );
-    }
-
-    #[test]
-    fn test_no_gravity_when_grounded() {
-        let mut world = World::new();
-        let (state, _temp_dir) = create_test_state();
-        world.insert_resource(state);
-
-        let entity = world
-            .spawn((
-                Velocity { vec: Vec3A::ZERO },
-                OnGround(true),
-                Position {
-                    coords: DVec3::new(0.0, 100.0, 0.0),
-                },
-                HasGravity,
-            ))
-            .id();
-
-        let mut schedule = Schedule::default();
-        schedule.add_systems(handle);
-
-        // Run the gravity system
-        schedule.run(&mut world);
-
-        let vel = world.get::<Velocity>(entity).unwrap();
-        assert_eq!(
-            vel.vec.y, 0.0,
-            "Velocity Y should remain zero when grounded"
-        );
-    }
-
-    #[test]
-    fn test_water_entity_gravity_not_in_water() {
-        let mut world = World::new();
-        let (state, _temp_dir) = create_test_state();
-        world.insert_resource(state);
-
-        let entity = world
-            .spawn((
-                Velocity { vec: Vec3A::ZERO },
-                OnGround(false),
-                Position {
-                    coords: DVec3::new(0.0, 100.0, 0.0),
-                },
-                HasGravity,
-                HasWaterDrag,
-            ))
-            .id();
-
-        let mut schedule = Schedule::default();
-        schedule.add_systems(handle);
-
-        // Run the gravity system
-        schedule.run(&mut world);
-
-        let vel = world.get::<Velocity>(entity).unwrap();
-        assert!(
-            vel.vec.y < 0.0,
-            "Water entity should have gravity applied when not in water"
-        );
-    }
-
-    #[test]
-    fn test_water_entity_no_gravity_when_grounded() {
-        let mut world = World::new();
-        let (state, _temp_dir) = create_test_state();
-        world.insert_resource(state);
-
-        let entity = world
-            .spawn((
-                Velocity { vec: Vec3A::ZERO },
-                OnGround(true),
-                Position {
-                    coords: DVec3::new(0.0, 100.0, 0.0),
-                },
-                HasGravity,
-                HasWaterDrag,
-            ))
-            .id();
-
-        let mut schedule = Schedule::default();
-        schedule.add_systems(handle);
-
-        // Run the gravity system
-        schedule.run(&mut world);
-
-        let vel = world.get::<Velocity>(entity).unwrap();
-        assert_eq!(
-            vel.vec.y, 0.0,
-            "Water entity should not have gravity when grounded"
-        );
-    }
-
-    #[test]
-    fn test_water_entity_in_water_no_gravity() {
-        let mut world = World::new();
-        let (state, _temp_dir) = create_test_state();
-
-        // Create a chunk with water blocks
-        let chunk_pos = ChunkPos::new(0, 0);
-        create_chunk_with_water(&state, chunk_pos);
-
-        world.insert_resource(state);
-
-        // Spawn entity at Y=65 (in water)
-        let entity = world
-            .spawn((
-                Velocity { vec: Vec3A::ZERO },
-                OnGround(false),
-                Position {
-                    coords: DVec3::new(0.0, 65.0, 0.0),
-                },
-                HasGravity,
-                HasWaterDrag,
-            ))
-            .id();
-
-        let mut schedule = Schedule::default();
-        schedule.add_systems(handle);
-
-        // Run the gravity system
-        schedule.run(&mut world);
-
-        let vel = world.get::<Velocity>(entity).unwrap();
-        assert_eq!(
-            vel.vec.y, 0.0,
-            "Water entity should not have gravity applied when in water (drag system handles it)"
-        );
-    }
-}
