+++ src/game_systems/tests/interactions.rs
@@ -1,0 +1,2 @@
+#[path = "interactions/block_interactions.rs"]
+mod block_interactions;
