+++ src/game_systems/tests/physics.rs
@@ -1,0 +1,4 @@
+#[path = "physics/gravity.rs"]
+mod gravity;
+#[path = "physics/velocity.rs"]
+mod velocity;
